==> poblacion_estuds/02 - Create training set from EPH.ipynb <==
## This notebook loads EPH data, cleans it and arranges it to be used as 'training sets'. 
That is, for fitting any Machine Learning model.

import warnings
warnings.filterwarnings('ignore')
import pandas as pd

import numpy as np
pd.options.display.max_columns = 99

import glob
path ='./../../Documents/EPH/microdatos/' # use your path

allFiles = glob.glob(path + 'hogar/*.txt')
frame = pd.DataFrame()
list_ = []
for file_ in allFiles:
    df = pd.read_csv(file_,index_col=None, header=0, delimiter = ';', 
                    usecols = ['CODUSU','ANO4','TRIMESTRE','IX_TOT', 'AGLOMERADO',
    'IV1', 'IV3', 'IV4','IV5','IV6','IV7','IV8','IV10','IV11','II1','II2','II7','II8','II9']) 
#     ['II2', 'IV5', 'IX_TOT', 'II7', 'IV4', 'II1', 'IV7', 'IV6', 'IV11', 'IV8', 'IV3', 'II8', 'IV1', 'IV10', 'II9']

    list_ += [df]
df = pd.concat(list_)

# Correcciones Respuestas. Para que matchee censo
df = df.loc[df.IV1 != 9]
df['IV10'] = df['IV10'].map({1: 1, 2: 2, 3: 2, 0: 0, 9: 9})
df['II9'] = df['II9'].map({1: 1, 2: 2, 3: 2, 4: 4, 0: 0})
df['II7'] = df['II7'].map({1: 1, 2: 2, 3: 3, 4: 4, 5: 5, 6: 6, 7: 6, 8: 6, 9: 6, 0: 0})
df['II9'] = df['II9'].map({1: 1, 2: 2, 3: 2, 4: 4, 0: 0})
df['IX_TOT'] = df['IX_TOT'].clip(0, 8)

hogar = df


allFiles = glob.glob(path + "individual/usu_individual*.txt")
frame = pd.DataFrame()
list_ = []
for file_ in allFiles:
    df = pd.read_csv(file_,index_col=None, header=0, delimiter = ';', 
                     usecols = ['CODUSU','ANO4','TRIMESTRE','CH04','CH06', 'AGLOMERADO', 'CH09','CH10','CH12','CH13','CH15'] +\
                     ['CH07','CAT_INAC','CAT_OCUP','PP07G1', 'PP07G2', 'PP07G3', 'PP07G4', 'PP07G_59', 'PP07H', 'PP07I', 'PP07J', 'PP07K',
                     'P47T', 'V3_M', 'T_VI', 'V12_M', 'TOT_P12', 'V5_M','V2_M', 'PP08D1', 'P21'])

# For the regression training set. But for these the ANO4 TRIMESTRE is important.. Also we need more memory.
#                      ['P21','P47T',,'CH08','CH16','TOT_P12','T_VI','V10_M','V11_M','V12_M','V18_M','V19_AM','V21_M','V2_M','V3_M',
#             'V4_M','V5_M','V8_M','V9_M','PP08D1','PP08D4','PP08F1','PP08F2','PP08J1','PP08J2','PP08J3','PP10A','PP10C','PP10D','PP10E'])

    list_ += [df]
df = pd.concat(list_)

# Correcciones Respuestas. Para que matchee censo
df['CH15'] = df['CH15'].map({1:1, 2:1, 3:1, 4:2, 5:2, 9:0})
df['CH06'] = df['CH06'].clip(0)
df['CH09'] = df['CH09'].map({1:1, 2:2, 0:2, 3:2})

indiv = df

EPH = hogar.merge(indiv[list(indiv.columns.difference(hogar.columns)) + ['CODUSU']], on = 'CODUSU')
EPH['CONDACT'] = (EPH['CH06'] >= 14) + 2 * (EPH['CAT_OCUP'] == 0)*(EPH['CH06'] >= 14)


###--------------------------------------------------------------------
### reducir EPH a los AGLOS presentes en el sample de censo elegido. 
### Ademas adosamos una copia con AGLO = 0, para todos los que no pertenecen a algun AGLO.
###--------------------------------------------------------------------

# Usar la EPH necesaria (i.e. sacar las provincias si no se van a usar)
# Solo se corre una vez!!
aglos_sel = np.array([32,  0, 33,  3,  2, 93, 38]) # CABA y PBA

if aglos_sel[0]:
    EPH_sel = EPH.loc[EPH['AGLOMERADO'].isin(aglos_sel)]
    
    EPH_no_aglo = EPH_sel.copy(); 
    EPH_no_aglo['AGLOMERADO'] = 0

    EPH_sel = pd.concat([EPH_sel, EPH_no_aglo]).reset_index(drop = True)
    
    EPH = EPH_sel

### Match column names

names_censo = ['IX_TOT', 'P02', 'P03', 'CONDACT', 'AGLOMERADO',
    'V01', 'H05', 'H06', 'H07', 'H08', 'H09', 'H10', 'H11', 'H12', 'H16', 'H15', 'PROP', 'H14', 'H13',
      'P07', 'P08', 'P09', 'P10', 'P05']


names_EPH = ['IX_TOT','CH04','CH06','CONDACT', 'AGLOMERADO',
    'IV1', 'IV3', 'IV4','IV5','IV6','IV7','IV8','IV10','IV11','II1','II2','II7','II8','II9',
    'CH09','CH10','CH12','CH13','CH15']


# Pool all years

PBA_training = EPH.drop(['ANO4', 'TRIMESTRE'], axis = 1).drop_duplicates()
PBA_training = PBA_training.rename(columns = dict(zip(names_EPH, names_censo)))
PBA_training.to_csv('PBA_train.csv', index = False)

PBA_training.shape

# Pool all years
for i, row in EPH[['ANO4', 'TRIMESTRE']].drop_duplicates().iterrows():
    Y, Q = row['ANO4'], row['TRIMESTRE']

    PBA_tT = EPH.loc[(EPH.ANO4 == Y) & (EPH.TRIMESTRE == Q)].drop_duplicates()
    PBA_training_TRIMESTRE = PBA_tT.rename(columns = dict(zip(names_EPH, names_censo)))
    PBA_training_TRIMESTRE.to_csv('PBA_train_'+str(Y)+str(Q)+'.csv', index = False)

EPH[['ANO4', 'TRIMESTRE']].drop_duplicates()

## Listo. Salvado el training set.

# Older code

### Filtrando censo a los DPTO elegidos

# import dask.dataframe as dd
# from dask.diagnostics import ProgressBar

# #Esto es para extraer las viviendas, hogares y personas de los partidos (DPTOs) en cuestion.

# VIVIENDA = dd.read_csv('./VIVIENDA.csv', sep = ';', usecols = ['VIVIENDA_REF_ID', 'RADIO_REF_ID', 'TIPVV', 'V01'])
# VIVIENDA = VIVIENDA.merge(radio_ref[['RADIO_REF_ID', 'DPTO']])
# VIVIENDA_ = VIVIENDA.loc[VIVIENDA.DPTO.isin(seleccion_DPTOS)]
# with ProgressBar():
#     VIVIENDA_REF_ID_sel = VIVIENDA_['VIVIENDA_REF_ID'].values.compute()

# HOGAR = dd.read_csv('./HOGAR.csv', sep = ';', usecols = ['HOGAR_REF_ID', 'VIVIENDA_REF_ID', 'H05', 'H06', 'H07', 'H08',
#        'H09', 'H10', 'H11', 'H12', 'H13', 'H14', 'H15', 'H16', 'PROP', 'TOTPERS']) # csv is too big, so it is dask-loaded. Not sure it's efficient thou
# # For example computing len takes ages
# # len(HOGAR.VIVIENDA_REF_ID)
# HOGAR_ = HOGAR.loc[HOGAR.VIVIENDA_REF_ID.isin(VIVIENDA_REF_ID_sel)]
# with ProgressBar():
#     HOGAR_REF_ID_sel = HOGAR_['HOGAR_REF_ID'].values.compute()

# PERSONA = dd.read_csv('./PERSONA.csv', sep = ';', usecols = ['PERSONA_REF_ID', 'HOGAR_REF_ID', 'P01', 'P02', 'P03', 'P05', 'P06',
#        'P07', 'P12', 'P08', 'P09', 'P10', 'CONDACT'])
# PERSONA_ = PERSONA.loc[PERSONA.HOGAR_REF_ID.isin(HOGAR_REF_ID_sel)]

# tabla_censo = VIVIENDA_.merge(HOGAR_)#.merge(PERSONA_)

# IX_TOT = tabla_censo.groupby('HOGAR_REF_ID').count().iloc[:, 0].reset_index()
# IX_TOT.columns = ['HOGAR_REF_ID', 'IX_TOT']

# tabla_censo = tabla_censo.merge(IX_TOT)

# with ProgressBar():
#     table = tabla_censo.compute()


### Making EPH and Censo answers uniform

# Approach: modify Census to fit EPH
# Armonizar para adecuar dataset Censo a las opciones rta de EPH. No correr 2 veces
table['V01'] = table['V01'].map({1:1, 2:6, 3:6, 4:2, 5:3, 6:4, 7:5, 8:6})
table['H06'] = table['H06'].map({1:1, 2:2, 3:3, 4:4, 5:5, 6:6, 7:7, 8:9})
table['H09'] = table['H09'].map({1:1, 2:2, 3:3, 4:4, 5:4, 6:4})
table['H16'] = table['H16'].clip(0, 9)
table['H14'] = table['H14'].map({1:1, 2:4, 3:2, 4:2, 5:4, 6:3, 7:4, 8:9})
table['H13'] = table['H13'].map({1:1, 2:2, 4:0})
# table['P07'] = table['P07'].map({1:1, 2:2, 0:2})

# saber de que aglo es la persona. Se usa los resultados de cada aglo.
table = table.merge(radio_ref[['RADIO_REF_ID','AGLOMERADO']]) 

# table = table.set_index('PERSONA_REF_ID')

# Only once to save time in the future
# name =  'bolivar'#''
# name =  'vlopez_rodriguez'#''
name =  'rand'#''
table.to_csv('./sample_censo_table'+str(n).zfill(3)+name+'.csv', index = False)

# table = pd.read_csv('./sample_censo_table'+str(n).zfill(3)+'.csv')#.csv


with ProgressBar():
    PERSONA_ = PERSONA_.compute()

table = table.merge(PERSONA_)

table['P07'] = table['P07'].map({1:1, 2:2, 0:2})

# Only once to save time in the future
table.to_csv('./sample_censo_table'+str(n).zfill(3)+name+'.csv', index = False)


# AGLOS_censo = 
radio_ref_sel = radio_ref.loc[table['RADIO_REF_ID'].drop_duplicates().values]
aglos_sel = radio_ref_sel.AGLOMERADO.unique()

## Entradas de censo de los DPTOs elegidos

print(table.shape) #cuanta (gente, variables)?
table.sample(5)

## Cargar EPH

### Adoptar mismos nombres de columnas

# Misma info, distinto nombre. 
# Censo INDEC 
md_1 = table[['IX_TOT', 'P02', 'P03', 'CONDACT', 'AGLOMERADO', #las que no se erran, cant pers, sexo, edad, act, aglo
    'V01', 'H05', 'H06', 'H07', 'H08', 'H09', 'H10', 'H11', 'H12', 'H16', 'H15', 'PROP', 'H14', 'H13',
      'P07', 'P08', 'P09', 'P10', 'P05']] #las x que buscan matches un poquito mas laxamente


#Mismas cosas, distinto nombre de columna para
# EPH INDEC
md_2 = EPH[['IX_TOT','CH04','CH06','CONDACT', 'AGLOMERADO',
    'IV1', 'IV3', 'IV4','IV5','IV6','IV7','IV8','IV10','IV11','II1','II2','II7','II8','II9',
    'CH09','CH10','CH12','CH13','CH15']]

# # Now we want to see in each column what are the percentages, as a clue to where there can be issues
# # OK control check. Control there is less likely confusion. 

# for i in range(len(md_1.columns))[:2]: 
#     print('\n')
#     for md in [md_1, md_2]:
#         col = md.columns[i]
#         print(col)
#         print(md[col].value_counts().sort_index()/len(md))

md_2.columns = md_1.columns

import numpy as np


### The 'y' variables will be predicted. K nearest neighbors is used.
## Variables in the EPH survey but not in the Censo.
## Preguntas de EncuestaPH que no estan en Censo.

y_cols = ['P21','P47T','CAT_INAC','CAT_OCUP','CH07','CH08','CH16','TOT_P12','T_VI','V10_M','V11_M','V12_M','V18_M','V19_AM','V21_M','V2_M','V3_M','V4_M','V5_M','V8_M','V9_M',
         'PP07G1', 'PP07G2', 'PP07G3', 'PP07G4', 'PP07G_59','PP07H','PP07I','PP07J','PP07K','PP08D1','PP08D4','PP08F1','PP08F2','PP08J1','PP08J2','PP08J3','PP10A','PP10C','PP10D','PP10E']

#Remove ill predicted ones after trying them out
y_cols = list(set(y_cols) - set(['PP10A', 'V11_M', 'PP08D4', 'PP08J3', 'PP08F1', 'V18_M', 'V10_M',
       'V8_M', 'V4_M', 'PP08F2', 'V21_M', 'V9_M', 'PP08J2', 'PP08J1',
       'V19_AM']))



train = md_2.join(EPH[y_cols]).fillna(0)
test = md_1.fillna(0).astype(int)


md_1.columns

train.to_csv('./train.csv', index = False)

len(train)

# Se entrena en .8% de la poblacion... con solo tener La Matanza y La Plata. Parece poco: error prone.
print(train.shape)
print(test.shape)
len(train)/len(test)

# Las variables principales que no se erran. Personas en hogar, su sexo, su decil de edad (~ grupo 6 anios) y condicion de actividad, o sea si sale a trabajar o 'esta en casa'.

# train[['IX_TOT', 'P02', 'P03', 'CONDACT']].nunique().values
# train[['IX_TOT', 'P02', 'P03', 'CONDACT']].drop_duplicates()

len(test[x_cols].drop_duplicates())/len(test[x_cols])
# test[x_cols]

# The columns that we will be making predictions with.
x_cols = md_1.columns
from sklearn.neighbors import KNeighborsRegressor
# Create the knn model.
# Look at the five closest neighbors.
knn = KNeighborsRegressor(n_neighbors=1)

# Fit the model on the training data.
train = train.sample(frac = 1)
#Add noise to reduce tie situations in a more or less random way
knn.fit(train[x_cols] + .1*(np.random.random_sample(train[x_cols].shape)-0.5), train[y_cols])

# %%timeit

# Make point predictions on the test set using the fit model.
predictions = knn.predict(test[x_cols])

pred_df = pd.DataFrame(predictions, columns=y_cols, index = test.index)

result = test.join(pred_df)

pd.options.display.max_columns = 990
pd.options.display.max_rows = 999

print(result.shape)
result.head()

result.to_csv('./test_result_'+str(n).zfill(3)+'dpto_1.csv')
# result.to_csv('./test_result_'+str(n).zfill(3)+'dpto.csv')

# Load result

res_1 = pd.read_csv('./test_result_'+str(n).zfill(3)+'dpto_1.csv')

# res_0 = pd.read_csv('./test_result_'+str(n).zfill(3)+'dpto.csv')

# porcentaje error por reshuffle: 
perr_ = np.round(100*abs(res_0.sum() - res_1.sum())/res_0.sum(), 1).sort_values()

# perr_.tail(15)#.index
# Bad ones: ['PP10A', 'V11_M', 'PP08D4', 'PP08J3', 'PP08F1', 'V18_M', 'V10_M',
#        'V8_M', 'V4_M', 'PP08F2', 'V21_M', 'V9_M', 'PP08J2', 'PP08J1',
#        'V19_AM']


import numpy as np
# Agregar $$$. En millones de usd
# En millones de usd (USD = 30 ARS)
_USD = 30.5 #ARS
np.round(res_1.sum()/_USD/1e6, 1).sort_values().tail(14)

#PPALES
# negocio que no trabajo no laborable (V9_M)
# alquiler no laborable (V8_M)
# indemnizacion despido no laborable (V3_M)
# comision Ocupacion ppal (PP08F1)
# cuota alimentos no laborable (V12_M)
# subsidio ayuda social no laborable (V5_M)
# TOTAL otras ocupacions(TOT_P12)
# jubilacion no laborable (V2_M)
# TOTAL no laborables (T_VI)
# sueldo Ocupacion ppal(PP08D1)
# TOTAL Ocupacion ppal (P21)
# TOTAL TOTAL (P47T)

# np.round(res_1.sum()/1e6/_USD, 1).sort_values().tail(14).index

PERS_DPTO = table[['PERSONA_REF_ID', 'RADIO_REF_ID']].merge(radio_ref_sel[['RADIO_REF_ID', 'DPTO' #, 'NOMDPTO', 'NOMPROV'
                                                                      ]]).drop(['RADIO_REF_ID'], axis = 1)

res = res_1
res_DPTO = res.merge(PERS_DPTO, on = 'PERSONA_REF_ID')

#en miles de USD
res_byDPTO = res_DPTO.groupby(['DPTO'])[[  'V3_M', 'V12_M', 'V5_M', 'TOT_P12',
       'V2_M', 'T_VI', 'PP08D1', 'P21', 'P47T']].sum()

np.round(100*res_byDPTO.div(res_byDPTO.P47T, axis = 0), 1).sort_values(by = 'P21').head() #percentage
# np.round(res_byDPTO/1e3/_USD, 1) #in 1000 USD

PERS_DPTO = table[['PERSONA_REF_ID', 'RADIO_REF_ID']].merge(radio_ref_sel[['RADIO_REF_ID', 'DPTO', 'NOMDPTO', 'NOMPROV']]
                                                           )#.drop(['RADIO_REF_ID'], axis = 1)

res = res_1
res_DPTO = res.merge(PERS_DPTO, on = 'PERSONA_REF_ID')

# variables = ['V9_M', 'V8_M', 'PP08F1', 'V3_M', 'V12_M', 'V5_M', 'TOT_P12',
#        'V2_M', 'T_VI', 'PP08D1', 'P21', 'P47T']
variables = [  'V3_M', 'V12_M', 'V5_M', 'TOT_P12',
       'V2_M', 'T_VI', 'PP08D1', 'P21', 'P47T']
#en miles de USD
# res_byDPTO = res_DPTO.groupby(['DPTO', 'NOMDPTO', 'NOMPROV'])[variables].sum()
res_byDPTO = res_DPTO.groupby(['RADIO_REF_ID'])[variables].sum()

# np.round(100*res_byDPTO.div(res_byDPTO.P47T, axis = 0), 1).sort_values(by = 'P21').head() #percentage
np.round(res_byDPTO/1e3/_USD, 1) #in 1000 USD

# Save info at 'radio' level
res_byDPTO.to_csv('res_byradio_sample_'+str(n).zfill(3)+'.csv')

# Desoc, NA = 0. Not good.
variables = ['PP07J', #turno habitual
 'PP10D', #Desoc. Ha trabajado alguna vez?
 'PP10C', #Desoc. Hizo changa mientras buscaba?
 'PP07K', # Oc. ppal. Inc. serv. dom. Cobra con recibo
 'PP07G2', # Oc. ppal. Inc. serv. dom. aguinaldo
 'PP07G_59', # Oc. ppal. Inc. serv. dom. ninguno
 'PP07G3', # Oc. ppal. Inc. serv. dom. dias enfermedad
 'PP10E', # Desoc. Tiempo de que termino su ultimo trabajo/changa
 'PP07H', # Oc. ppal. Inc. serv. dom. descuento jubilatorio
 'PP07G4', # Oc. ppal. Inc. serv. dom. obra social
 'PP07I', # Oc. ppal. Inc. serv. dom. Aporta jub por sí mismo 
 'PP07G1', # Oc. ppal. Inc. serv. dom. vacaciones pagas
 'CH07'] #Est civil

#en miles de USD
# res_byDPTO = res_DPTO.groupby(['DPTO', 'NOMDPTO', 'NOMPROV'])[variables].mean()
res_byDPTO = res_DPTO.groupby(['RADIO_REF_ID'])[variables].mean()


s = np.round(res_byDPTO, 2).sort_values(by = 'PP07K')#.head() 
s.style.bar(color='#d65f5f')
#  'CAT_OCUP', #CAT_INAC
#  'CAT_INAC', #CAT_INAC

#  'CH08', obra social/salud. nums altos como para mean
# 'CH16', # Donde vivia hace 5. Desconfiable


# from sklearn.neighbors import NearestNeighbors
# neigh = NearestNeighbors(n_neighbors=1)
# neigh.fit(train[x_cols], train.sample(frac = 1)[y_cols])
# i = 60
# print(neigh.kneighbors([test.iloc[i].values], return_distance=True))
# train.iloc[7022][x_cols] - test.iloc[i]

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

for col in result.columns:
    print('\n')
    print(col)
    df_ = result.loc[result.P03 > 2]
    print(df_[col].value_counts().sort_index())
